Улучшил работу с временными метками

+++ feature_main/src/main/java/com/learning/lessons/feature_main/ui/lesson/TimecodesAdapter.kt
@@ -27,7 +27,7 @@ class TimecodesAdapter(val data: List<Timecode>, val  onItemClickListener: (Long
                 tvTime.text = data.time
                 tvTibeDiscribe.text = data.title
                 this.root.setOnClickListener {
-                    onItemClickListener(data.timeSeconds)
+                    onItemClickListener(data.timeSeconds.toLong())
                 }
             }
         }
